requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
python-multipart
google-generativeai
streamlit
